Provide access to stack traces of arbitrary size in extended data items.
The current StackTrace32(_) and StackTrace64(_) definitions yield a value of the inner item type exposed by the windows crate. This type is autogenerated from the C definition, which uses an ANYSIZE_ARRAY (a constant defined as 1). As a side effect, a value of the autogenerated Rust type can only contain a single address from the captured call stack.

* Define a custom type usable as StackTraceItem<u32> and StackTraceItem<u64> to represent captured call stacks of arbitrary size.

* Make StackTrace32(_) and StackTrace64(_) yield values of this new custom type.