Make get_provider_guid use the TDH API
The get_provider_guid function was built on top of the PLA COM
interface. This was quite easy to use but had the drawback of needing
to call CoInitializeEx (which had side effects such as being slow, and
indirectly leaking memory
(https://github.com/microsoft/windows-rs/issues/1169 for the rabbit
hole)).

This commit mimics what's done in KrabsETW and uses the TDH API to
implement the function.